vfs: consistently use size_t for buflen around VOP_VPTOCNP
git-svn-id: svn+ssh://svn.freebsd.org/base/head@357383 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/kern/vfs_default.c
@@ -804,7 +804,7 @@ vop_stdvptocnp(struct vop_vptocnp_args *ap)
 	struct vnode **dvp = ap->a_vpp;
 	struct ucred *cred = ap->a_cred;
 	char *buf = ap->a_buf;
-	int *buflen = ap->a_buflen;
+	size_t *buflen = ap->a_buflen;
 	char *dirbuf, *cpos;
 	int i, error, eofflag, dirbuflen, flags, locked, len, covered;
 	off_t off;
